Implementanción de roles a los enpoints, se carga el rol desde la base de datos y se valida para luego adjuntarlo en el token.

+++ domain/model/src/main/java/co/com/pragma/model/solicitante/Solicitante.java
@@ -20,7 +20,7 @@ public class Solicitante {
     private String documentoIdentidad;
     private LocalDate fechaNacimiento;
     private String clave;
-    private int idRol;
+    private Long idRol;
     private String direccion;
     private String telefono;
     private String correoElectronico;
