+++ infrastructure/driven-adapters/r2dbc-postgresql/src/main/java/co/com/pragma/r2dbc/entity/SolicitanteEntity.java
@@ -17,7 +17,7 @@ public class SolicitanteEntity {
     private String documentoIdentidad;
     private LocalDate fechaNacimiento;
     private String clave;
-    private int idRol;
+    private Long idRol;
     private String direccion;
     private String telefono;
     private String correoElectronico;
