+++ infrastructure/entry-points/reactive-web/src/main/java/co/com/pragma/api/dto/SolicitanteRequest.java
@@ -21,12 +21,11 @@ public class SolicitanteRequest {
     @Past(message = "La fecha de nacimiento debe ser una fecha pasada")
     private LocalDate fechaNacimiento;
 
-
     @NotNull
     private String clave;
 
     @NotNull
-    private int idRol;
+    private Long idRol;
 
     @NotBlank
     private String direccion;
